chore(hangman-game): rename utile.js to util.js
