+++ hangman-game/index.html
@@ -8,6 +8,8 @@
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@400;700&display=swap" rel="stylesheet">
+  <link rel="stylesheet" href="./src/index.css">
+  <link rel="stylesheet" href="./src/app.css">
 </head>
 
 <body>
@@ -28,7 +30,7 @@
     <code id="state" style="display:none"></code>
   </div>
 
-  <script src="./src/index.js"></script>
+  <script type="module" src="./src/index.js"></script>
 </body>
 
 </html>
