+++ hangman-game/src/index.js
@@ -1,5 +1,5 @@
 import App from "./app";
-import "./index.css";
+
 
 const run = () => {
   window.addEventListener("DOMContentLoaded", () => {
